fix more warnings

+++ ini.h
@@ -44,8 +44,8 @@ static bool parse_line(char* line, pair* c) {
   uint32_t key_start = 0; uint32_t value_start = 0;
   bool start_value = false; bool start_key = false;
   uint32_t prev_space = 0;
-  for (int32_t i = 0; i < len - 1; i++) {
-    if (i < equal_index) {
+  for (uint32_t i = 0; i < len - 1; i++) {
+    if (i < (uint32_t)equal_index) {
       if (line[i] == ' ' && !prev_space) {
         prev_space = key_start;
       } else if (line[i] != ' ') {
@@ -54,7 +54,7 @@ static bool parse_line(char* line, pair* c) {
       }
       if (start_key) 
         c->key[key_start++] = line[i];
-    } else if (i > equal_index) {
+    } else if (i > (uint32_t)equal_index) {
       if (line[i] != ' ') 
         start_value = true;
       if (start_value)
